sms-modem/requirements.txt
Flask
pika
PyYAML